refactor(wallet): merge validation errors into errors
