+++ wallet/src/actors/app/error.rs
@@ -4,7 +4,6 @@ use serde_json::json;
 
 use witnet_net::client::tcp;
 
-use super::*;
 use crate::actors;
 
 #[derive(Debug, Fail)]
@@ -95,5 +94,33 @@ impl From<actors::worker::Error> for Error {
 impl From<tcp::Error> for Error {
     fn from(err: tcp::Error) -> Self {
         node_error(err)
+    }
+}
+
+/// A list of errors. An error is a pair of (field, error msg).
+pub type ValidationErrors = Vec<(String, String)>;
+
+/// Create an error message associated to a field name.
+pub fn field_error<F: ToString, M: ToString>(field: F, msg: M) -> ValidationErrors {
+    vec![(field.to_string(), msg.to_string())]
+}
+
+/// Combine two Results but accumulate their errors.
+pub fn combine_field_errors<A, B, C, F>(
+    res1: std::result::Result<A, ValidationErrors>,
+    res2: std::result::Result<B, ValidationErrors>,
+    combinator: F,
+) -> std::result::Result<C, ValidationErrors>
+where
+    F: FnOnce(A, B) -> C,
+{
+    match (res1, res2) {
+        (Err(mut err1), Err(err2)) => {
+            err1.extend(err2);
+            Err(err1)
+        }
+        (Err(err1), _) => Err(err1),
+        (_, Err(err2)) => Err(err2),
+        (Ok(a), Ok(b)) => Ok(combinator(a, b)),
     }
 }
